Add a builder API for constructing KMS MKPs
The current KMS MKP constructors use legacy (@Deprecated) KMS client
constructors, and are likely to break at some point in the future. In addition,
they also mutate the client that is passed in, and have an unfortunate
combinatorial explosion of argument types.

This change introduces a new builder API for constructing KMS MKPs, and
deprecates the old one. In addition, it adds support for decrypting KMS keys
from multiple regions with the same MKP, bringing us to feature parity with the
Python SDK.

For now, the semantics of code using the old constructors is unchanged, but
it's likely that we'll want to remove these constructors the next time we make
a breaking change to our APIs.